Fix issues with atomics and selective coherence
This commit fixes a segfault issue in the ArgoDSM atomics, which
also extended to the ArgoDSM global locks. It also fixes a
stalling issue with selective coherence.
Finally, the default ARGO_MPI_WIN_GRANULARITY was increased from
16 to 64 to allow running of make test and smaller test programs
on most platforms. 16 may still be a useful size on systems with
more available memory.